added special buttons, added handlers, added temporary padding for styling, added navigate paths to buttons
