+++ src/Templates/ColumnWrapper/ColumnWrapperStyled.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-const ColumnWrapperStyled = styled.form`
+const ColumnWrapperStyled = styled.div`
   display: flex;
   width: 100%;
   flex-direction: column;
